Check for folders existence before deleting them from git

+++ upload_sources.sh
@@ -32,7 +32,9 @@ createcommit(){
   if [ "$sdkversion" -le "$lowestapi" ]; then
     for i in $(seq 1 "$((sdkversion - 1))")
     do
-      git rm -q -r --ignore-unmatch "$SOURCES/$2/$type/$package/$i/$dpis"
+      if [ "$(git ls-tree -r --name-only master "$type/$package/$i")" != "" ]; then
+        git rm -q -r --ignore-unmatch "$SOURCES/$2/$type/$package/$i/$dpis"
+      fi
     done
   fi
   # We don't have to care about empty direcories with git (see http://stackoverflow.com/a/10075480/3315861 for more details.)
